// src/users/generated/user-client.ts
export interface RoleResponse {
    id: string;
    name: string;
}

export interface AppResponse {
    appId: string;
    appName: string;
}

export interface AuthResponse {
    application: AppResponse;
    created: string;
}

export interface UserResponse {
    id: string;
    username: string;
    firstName: string;
    lastName: string;
    email: string;
    created: Date;
    hasGrantedAccess: boolean;
    grantedAccessOn: Date;
    roles: RoleResponse[];
    apps: AppResponse[];
    auths: AuthResponse[];
}